[WIP] first run

+++ src/entities/game.rs
@@ -51,7 +51,7 @@ pub struct Game {
     pub added_datetime: NaiveDateTime,
     pub updated_datetime: NaiveDateTime,
     pub status: i16,
-    pub rating: i32,
+    pub rating: i16,
     pub notes: String,
 }
